# openapi.yaml
openapi: 3.0.3
info:
  title: Helix Health Portal API
  description: |
    RESTful API for the Helix Health Portal outpatient patient portal.
    Used as the sample application for the QA automation course.
  version: 1.0.0
  contact:
    name: Helix Health Portal
servers:
  - url: http://localhost:4000/api
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiSuccess:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid credentials"
    AuthUser:
      type: object
      required: [id, email, roles]
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: patient1@helixhealthportal.test
        roles:
          type: array
          items:
            type: string
          example: ["patient"]
    LoginSuccessData:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/AuthUser'
    LoginMfaData:
      type: object
      required: [mfaRequired, mfaChallengeToken]
      properties:
        mfaRequired:
          type: boolean
          example: true
        mfaChallengeToken:
          type: string
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new patient account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: new.patient@example.com
                password:
                  type: string
                  minLength: 8
                  example: TestPass123!
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                          userId:
                            type: integer
                          mrn:
                            type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: patient1@helixhealthportal.test
                password:
                  type: string
                  example: TestPass123!
      responses:
        '200':
          description: Login successful (or MFA required)
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: '#/components/schemas/ApiSuccess'
                      - type: object
                        properties:
                          data:
                            $ref: '#/components/schemas/LoginSuccessData'
                  - allOf:
                      - $ref: '#/components/schemas/ApiSuccess'
                      - type: object
                        properties:
                          data:
                            $ref: '#/components/schemas/LoginMfaData'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Account inactive or email not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '429':
          description: Account locked (too many failed attempts)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token and log out
      operationId: logout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/mfa/setup:
    post:
      tags: [Auth]
      summary: Generate a TOTP secret and QR code for MFA setup
      operationId: mfaSetup
      security:
        - bearerAuth: []
      responses:
        '200':
          description: TOTP secret and QR code
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          secret:
                            type: string
                          otpauthUri:
                            type: string
                          qrCodeDataUrl:
                            type: string
  /auth/mfa/verify:
    post:
      tags: [Auth]
      summary: Verify a TOTP code and enable MFA on the account
      operationId: mfaVerify
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  pattern: '^\d{6}$'
                  example: '123456'
      responses:
        '200':
          description: MFA enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '400':
          description: Invalid TOTP code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/mfa/validate:
    post:
      tags: [Auth]
      summary: Complete login for MFA-enabled accounts (second step)
      operationId: mfaValidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mfaChallengeToken, code]
              properties:
                mfaChallengeToken:
                  type: string
                code:
                  type: string
                  pattern: '^\d{6}$'
                  example: '123456'
      responses:
        '200':
          description: Login complete
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginSuccessData'
        '401':
          description: Invalid code or expired challenge token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/password-reset/request:
    post:
      tags: [Auth]
      summary: Request a password-reset link (mocked — logged to console)
      operationId: passwordResetRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset link sent (always 200 to prevent user enumeration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Confirm password reset with the token from the email link
      operationId: passwordResetConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
                  example: NewTestPass456!
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '400':
          description: Validation error (e.g., password complexity)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Invalid or expired reset token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # ── Patients ─────────────────────────────────────────────────────────────────
...